+++ graphix/circ_ext/extraction.py
@@ -67,7 +67,6 @@ class ExtractionResult:
         Circuit
             Quantum circuit represented as a set of instructions.
         """
-        # Circumvent import loop
         from graphix.circ_ext.compilation import er_to_circuit  # noqa: PLC0415
 
         return er_to_circuit(self, pexp_cp=pexp_cp, cm_cp=cm_cp)