+++ graphix/opengraph.py
@@ -20,10 +20,9 @@ if TYPE_CHECKING:
     # Unpack introduced in Python 3.12
     from typing_extensions import Unpack
 
+    from graphix import CausalFlow, Pattern
     from graphix.circ_ext.extraction import CliffordMap, PauliExponentialDAG
-    from graphix.flow.core import CausalFlow
     from graphix.parameter import ExpressionOrSupportsFloat, Parameter
-    from graphix.pattern import Pattern
     from graphix.transpiler import Circuit
     from graphix.visualization import DrawKwargs