+++ noxfile.py
@@ -114,10 +114,10 @@ class ReverseDependency:
         ),
         ReverseDependency("https://github.com/TeamGraphix/graphix-qasm-parser", branch="fix_angles"),
         ReverseDependency(
-            "https://github.com/emlynsg/veriphix",
-            branch="jcz-test-fix",
+            "https://github.com/thierry-martinez/veriphix",
             doctest_modules=False,
             install_target=".[dev]",
+            branch="fix/graphix_namespace",
         ),
         ReverseDependency("https://github.com/TeamGraphix/graphix-ibmq", doctest_modules=False),
         ReverseDependency("https://github.com/qat-inria/graphix-stim-compiler", branch="ps_dim"),